[ALPS08631203] fix build error due to mainline migration
[Description]
Fix build errors associated with merging Linux 6.11 into the android-mainline.

[Test]
Build and Boot up

MTK-Commit-Id: 2ec28a8637aa1a4a1e3b6cae1570a72dbb22f3b2

CR-Id: ALPS08631203
Feature: Kernel SI Operation

+++ drivers/dma-buf/heaps/mtk_sec_heap.c
@@ -1709,8 +1709,8 @@ static struct dma_buf *alloc_dmabuf(struct dma_heap *heap,
 
 static struct dma_buf *tmem_page_allocate(struct dma_heap *heap,
 					  unsigned long len,
-					  unsigned long fd_flags,
-					  unsigned long heap_flags)
+					  u32 fd_flags,
+					  u64 heap_flags)
 {
 	int ret = -ENOMEM;
 	struct dma_buf *dmabuf;
@@ -1786,8 +1786,8 @@ free_buffer:
 
 static struct dma_buf *tmem_region_allocate(struct dma_heap *heap,
 					    unsigned long len,
-					    unsigned long fd_flags,
-					    unsigned long heap_flags)
+					    u32 fd_flags,
+					    u64 heap_flags)
 {
 	int ret = -ENOMEM;
 	struct dma_buf *dmabuf;
